נסיונות/nimas/server/requirements.txt
Flask==2.0.1
pymongo==4.8.0
requests==2.32.3
autopep8==2.3.1
celery==5.4.0
python-dotenv==1.0.1
tasks==2.8.0